week7_day1_hw/Single_Responsibility/app/src/main/java/example/codeclan/com/singleresponsibility/Journal.java
package example.codeclan.com.singleresponsibility;

import java.util.ArrayList;

/**
 * Created by kynansong on 06/11/2017.
 */

public class Journal {
    private ArrayList<String> journalEntries;

    public Journal() {
        journalEntries = new ArrayList<String>();
    }

    public void writeJournal(String entry){
        journalEntries.add(entry);
    }

    public String getJournalEntry(int index){
        return journalEntries.get(index);
    }

    public String getLastJournalEntry(){
        return getJournalEntry(journalEntries.size() - 1);
    }


}
